Add files via upload
Current project is all about the analysis of sales and consumer buying behavior of  superstore sales. Project deals with the in depth data of 51291 customers across the world provided from 2011 till 2015. This includes planning, Data collection, Organization of information, Analysis of information collected, Interpretation and disclosure in a clear and objective manner.Project workspace includes :
1)Tables ( serves to organize and tabulate the data) and
2)Graphs (Convey the information with clarity and transparency, contributing to an objective reading)Expected output - Solutions for the questioned problem statements with multi dimensional visualization of the provided data.